finishFrontEnd

+++ front_end/desafio-full-stack/src/contexts/AuthContext.tsx
@@ -2,12 +2,13 @@ import { Api } from "../services/api";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
 import React, { useState, createContext, useEffect } from "react";
+import jwtDecode from "jwt-decode";
 
 interface IAuthContextProps {
   children: React.ReactNode;
 }
 
-interface IContactsInfo {
+export interface IContactsInfo {
   id: string;
   fullName: string;
   telephone: string;
@@ -15,7 +16,7 @@ interface IContactsInfo {
   createdAt: string;
 }
 
-interface IClientInfo {
+interface IUserInfo {
   id: string;
   fullName: string;
   telephone: string;
@@ -29,10 +30,27 @@ export interface ILogin {
   password: string;
 }
 
+export interface ISignUp {
+  fullName: string;
+  email: string;
+  password: string;
+  confirmPass: string;
+  telephone: string;
+}
+
 interface IAuthContext {
-  clientInfo: IClientInfo;
+  userInfo: IUserInfo;
   token: string | null;
+  userId: string | null;
   login: (user: ILogin) => void;
+  signUp: (user: ISignUp) => void;
+  registerContact: (contact: IContactsInfo) => void;
+  showModalRegister: boolean;
+  setShowModalRegister: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+interface IDecodedToken {
+  sub: string;
 }
 
 export const AuthContext = createContext<IAuthContext>({} as IAuthContext);
@@ -40,14 +58,19 @@ export const AuthContext = createContext<IAuthContext>({} as IAuthContext);
 export const AuthProvider = ({ children }: IAuthContextProps) => {
   const navigate = useNavigate();
   const token: string | null = localStorage.getItem("@desafioFullStack:Token");
+  const userId: string | null = localStorage.getItem("@desafioFullStack:ID");
 
-  const [clientInfo, setClientInfo] = useState<IClientInfo>({} as IClientInfo);
+  const [userInfo, setUserInfo] = useState<IUserInfo>({} as IUserInfo);
+  const [showModalRegister, setShowModalRegister] = useState<boolean>(false);
 
   const login = (user: ILogin) => {
     Api.post("/users/login", { ...user })
       .then((res) => {
         window.localStorage.clear();
         window.localStorage.setItem("@desafioFullStack:Token", res.data.token);
+        const decodedToken = jwtDecode(res.data.token) as IDecodedToken;
+        const userId = decodedToken.sub;
+        window.localStorage.setItem("@desafioFullStack:ID", userId);
         toast.success("Login bem sucedido!");
         navigate("/dashboard", { replace: true });
       })
@@ -56,22 +79,58 @@ export const AuthProvider = ({ children }: IAuthContextProps) => {
       });
   };
 
+  const signUp = (user: ISignUp) => {
+    Api.post("/users", { ...user })
+      .then((res) => {
+        toast.success("Cadastro efetuado com sucesso");
+        setShowModalRegister(false);
+      })
+      .catch((err) => {
+        toast.error("E-mail já cadastrado!");
+      });
+  };
+
+  const registerContact = (contact: IContactsInfo) => {
+    Api.post(
+      "/contacts",
+      { ...contact },
+      { headers: { Authorization: `Bearer ${token}` } }
+    )
+      .then((res) => {
+        toast.success("Contato registrado com sucesso!");
+      })
+      .catch((err) => {
+        toast.error("Algo deu errado, tente novamente!");
+      });
+  };
+
   useEffect(() => {
     token &&
-      Api.get<IClientInfo>("/clients", {
+      Api.get<IUserInfo>(`/users/${userId}`, {
         headers: { Authorization: `Bearer ${token}` },
       })
         .then((res) => {
-          setClientInfo(res.data);
+          setUserInfo(res.data);
         })
         .catch((err) => {
           window.localStorage.clear();
           navigate("/login");
         });
-  }, [token, navigate, clientInfo.contacts]);
+  }, [token, navigate, userInfo.contacts, userId]);
 
   return (
-    <AuthContext.Provider value={{ token, clientInfo, login }}>
+    <AuthContext.Provider
+      value={{
+        token,
+        userId,
+        userInfo,
+        login,
+        signUp,
+        registerContact,
+        setShowModalRegister,
+        showModalRegister,
+      }}
+    >
       {children}
     </AuthContext.Provider>
   );
